blouch: Bayesian Linear Ornstein-Uhlenbeck models for Comparative Hypotheses

Blouch: Bayesian Linear Ornstein-Uhlenbeck models for Comparative Hypotheses fits adaptive models of continuous trait evolution in a Bayesian framework based on categorical or continuous predictors, incorporates measurement error. Blouch includes new implementations of Ornstein-Ulenbeck models including allowing for varying effects (varying intercepts and varying slopes), multilevel modeling, and non-centered models.

A DESCRIPTION OF THE PACKAGE

References

Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org

See also

Author

Maintainer: Mark Grabowski mark.walter.grabowski@gmail.com (ORCID)