blouch.adapt.prep - Setup dat file to run adaptive model

blouch.adapt.prep - Setup dat file to run adaptive model

blouch.adapt.prep(
  trdata,
  Y,
  Y_error,
  X,
  X_error,
  hl.prior,
  vy.prior,
  optima.prior,
  beta.prior
)

Arguments

trdata

An object of the class treedata from function treeplyr

Y

Vector containing name of column in treedata containing response variable

Y_error

Vector containing name of column in treedata containing error of response variable

X

Vector containing name(s) of column in treedata containing predictor variable(s)

X_error

Vector containing name(s) of column in treedata containing error of predictor variable(s)

Z_adaptive

Number of adaptive traits

Value

dat - list file containing objecs setup for Blouch