blouch.reg.direct.adapt.prep - Setup dat file to run multi-optima direct effect adaptive model

blouch.reg.direct.adapt.prep - Setup dat file to run multi-optima direct effect adaptive model

Usage

blouch.reg.direct.adapt.prep(
  trdata,
  Y,
  Y_error,
  X,
  X_error,
  Z_direct,
  Z_adaptive,
  reg.column,
  hl.prior,
  vy.prior,
  optima.prior,
  beta.prior
)

Arguments

trdata

An object of the class treedata from function treeplyr

Y

Vector containing name of column in treedata containing response variable

Y_error

Vector containing name of column in treedata containing error of response variable

X

Vector containing name(s) of column in treedata containing predictor variable(s)

X_error

Vector containing name(s) of column in treedata containing error of predictor variable(s)

Z_direct

Vector containing number of direct effect predictor traits

Z_adaptive

Vector containing number of adaptive predictor traits

reg.column

Vector containing name of regime column in treedata$dat

hl.prior

Vector containing log mu and standard error for half-life prior

vy.prior

Numeric with prior for scale of Vy

optima.prior

Vector containing mu and standard error for optima prior

beta.prior

Vector containing mu and standard error for beta prior

Value

dat - list file containing objecs setup for Blouch